requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.2.1
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.3.1
cohere==5.20.0
coloredlogs==15.0.1
cryptography==46.0.3
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.124.4
fastavro==1.12.1
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface_hub==1.2.3
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==34.1.0
langchain-core==1.2.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.11.0
openpyxl==3.1.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
ormsgpack==1.12.0
overrides==7.7.0
packaging==25.0
pandas==2.1.4
passlib==1.7.4
platformdirs==4.5.1
posthog==7.0.1
pre_commit==4.5.0
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.9
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
Pygments==2.19.2
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sse-starlette==2.1.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
ty==0.0.1a34
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.24.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
